make Keywords and TypeAnn classes unique
